[Embedding] Reduce the redundant code of EmbeddingVariable.

+++ tensorflow/core/framework/embedding/value_ptr.h
@@ -132,7 +132,7 @@ struct NormalHeader {
         freq_counter, freq_counter + 1);
   }
 
-  inline void AddFreq(int64 count) {
+  inline void AddFreq(int count) {
     __sync_bool_compare_and_swap(&freq_counter,
         freq_counter, freq_counter + count);
   }
@@ -178,7 +178,7 @@ struct FixedLengthHeader {
         freq_counter, freq_counter + 1);
   }
 
-  inline void AddFreq(int64 count) {
+  inline void AddFreq(int count) {
     __sync_bool_compare_and_swap(&freq_counter,
         freq_counter, freq_counter + count);
   }
@@ -273,7 +273,7 @@ class ValuePtr {
     LOG(FATAL) << "Unsupport FreqCounter in subclass of ValuePtrBase";
   }
 
-  virtual void AddFreq(int64 count) {
+  virtual void AddFreq(int count) {
     LOG(FATAL) << "Unsupport FreqCounter in subclass of ValuePtrBase";
   }
 
@@ -338,7 +338,7 @@ class NormalValuePtr : public ValuePtr<V> {
     return ((NormalHeader*)this->ptr_)->AddFreq();
   }
 
-  void AddFreq(int64 count) {
+  void AddFreq(int count) {
     MetaHeader* meta = (MetaHeader*)this->ptr_;
     return ((NormalHeader*)this->ptr_)->AddFreq(count);
   }
@@ -409,7 +409,7 @@ class NormalContiguousValuePtr : public ValuePtr<V>{
     ((FixedLengthHeader*)this->ptr_)->AddFreq();
   }
 
-  void AddFreq(int64 count) {
+  void AddFreq(int count) {
     ((FixedLengthHeader*)this->ptr_)->AddFreq(count);
   }
